Add functionality shot Do Not/Save
The save and do not save buttons have been added which delay the pulse
increase and save subVIs. As expected, the do not save button simply does
not save, clears the parameters, and resets the shot.

The functionality is not perfect, as if one does not save, then attempts
to save with the next shot, an error is thrown, though not a
program-killing error. To be fixed asap.