Add functionality to select different tree
Introduced a combo box that operates within the initialize stage. Used a
dreaded 'Local Variable' to transmit the tree string to the only other
location that the tree name is called; located within ShotLogSave.

Note: currently have to choose tree prior to lauching the VI. Might be
worthwhile to add some sort of 'Update' button in case the user wishes to
switch trees once the program has loaded.