Fix unequal DAQ channels bug. Append fix elsewhere
Here we have fixed the bug where an error is thrown when unequal nubers of
DAQ channels are active. The fix was to create a variable that defines the
actual active number of channels per DAQ card. THis fix was appended to
the entire DigitizerStorage VI, in that this subVI is bug-free.*

*100% sure the VI is entirely NOT bug-free.