Add prototype 'Load Previous Shot Settings' funct
Note that it is not currently operational! The framework is in place, and
the prototype has been implemented. THe prototype uses tags, which have
not been implemented yet in the MOCHImodel tree. There will be further
errors with various items, but as of now it does not break the code. To
reiterate, the Load ExpState VI does not impede the
Experiment_Control.VI!