Complete Prototype of MDS load previous shot
This is somewhat of a milestone. In this commit, the tree is all tagged,
and the MDSplus storage driver has been updated to reflect this, and has
been tested working. The MDSplus load does not throw any errors, but has
bugs in the code (not working correctly). Just wait on clicking that 'load
previous shot' button. It's tempting, I know, but it will mess up your
current settings. Feel free to experiment, by all means!