Fix MDS+ load bugs. Also modify timing diagram.
The MDS+ load works correctly. The timing diagram was modified such that
channels that are off and have no string length do not appear on the
diagram itself. In MDS+load, there is an error loading the Card-wide end
time. This was traced back to the actual initialization of the card-wide
End Time itself. The fix will be to fomulate the proper VI.