Add new updated firmware
Also various bug fixes that were done "on-the-fly." What I mean is that my
version of the code was under deep rennovation for the bank charging
simulation, and there was a small bug that needed to be fixed, and a new
version of MOCHIcontrol needed to be rolled out.

Also added a Bitfile folder structure for current/old versions.